+++ CMakeLists.txt
@@ -1,10 +1,10 @@
 # Uncompilable CMake File to enable project editing with CLion IDE
 
-cmake_minimum_required(VERSION 2.8.4)
+cmake_minimum_required(VERSION 3.13)
 project(arduino-mqtt)
 
 include_directories(
-        /Applications/Arduino.app/Contents/Java/hardware/arduino/avr/cores/arduino/
+        /Applications/Arduino.app/Contents/Java/hardware/arduino/avr/cores/arduino
         /Users/256dpi/Development/Arduino/libraries/Ethernet/src
         /Users/256dpi/Development/Arduino/libraries/WiFi101/src
         /Users/256dpi/Development/Arduino/libraries/MKRGSM/src
@@ -12,11 +12,11 @@ include_directories(
         /Users/256dpi/Library/Arduino15/packages/esp8266/hardware/esp8266/2.3.0/libraries/ESP8266WiFi/src
         /Users/256dpi/Library/Arduino15/packages/esp32/libraries/WiFi/src
         /Users/256dpi/Library/Arduino15/packages/esp32/libraries/WiFiClientSecure/src
-        src/)
+        src)
 
 include_directories(src/)
 
-set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++11")
+set(CMAKE_CXX_FLAGS -std=c++11)
 
 set(SOURCE_FILES
     examples/AdafruitHuzzahESP8266/AdafruitHuzzahESP8266.ino
@@ -33,6 +33,7 @@ set(SOURCE_FILES
         examples/ESP32DevelopmentBoardSecure/ESP32DevelopmentBoardSecure.ino
     src/lwmqtt
     src/MQTT.h
-    src/MQTTClient.h)
+    src/MQTTClient.h
+    src/MQTTClient.cpp)
 
 add_executable(arduino-mqtt ${SOURCE_FILES})
